Added more integration tests in IntegrationTests.java (Criticals)

+++ src/test/java/com/example/inventory_management/InventoryManagementApplicationTests.java
@@ -8,8 +8,10 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -29,7 +31,7 @@ public class InventoryManagementApplicationTests {
     @BeforeEach
     void setup() {
         prod.setName("Laptop");
-        prod.setPrice(10.00F);
+        prod.setPrice(BigDecimal.valueOf(10));
     }
 
     @Test
